refactor: separate post list
