+++ src/components/social/PostView.tsx
@@ -6,7 +6,7 @@ import {
 
 import Image from "next/image";
 import Link from "next/link";
-import { useState, useRef, useEffect, use } from "react";
+import { useState, useRef, useEffect } from "react";
 
 interface CommentDataType {
   photoUrl: string;
