+++ src/pages/test/social/index.tsx
@@ -1,5 +1,5 @@
 import PageLayout from "@/containers/social/PageLayout";
-import SocialPostList from "@/components/social/SocialPostList";
+import SocialPostList from "@/containers/social/PostList/List";
 
 import type { ReactElement } from "react";
 import type { NextPageWithLayout } from "../../_app";
